chore: Added @otaviomacedo and @HBobertz as code owners. (#106)
This repository has two "main" branches: `v2-main` and `main`, used for major version 2 and 3, respectively. Version 3 is still a release candidate, and it will be functionally identical to version 2, except that it will depend on the AWS SDK v3 (which implies some breaking API changes).

While the migration to AWS SDK v3 is under way, we have two options to achieve feature parity between the versions of `cdk-assets`:

1. Enforce all PRs to be duplicated across the branches, which is difficult to do.
2. Restrict the changes to only the `v2-main` branch, and, before releasing version 3, copying over all the code from `v2-main` to `main`.

This PR defines Hogan Bobertz and myself as code owners for most of the files in this repository. This, together with the branch protection rules, ensures that we can gatekeep any PRs until we are done with the migration.