feat: detect foreign bucket ownership (#111)
Currently, we don't differentiate between when:

1) We don't have access to the bootstrap bucket
2) We have access, but the bootstrap bucket belongs to a foreign account

It makes it so we treat (2) as a valid bucket ownership. This may or may not be desirable. 
This PR adds the aforementioned differentiation, allowing us to treat foreign bucket ownership is a special way.   

The trick is to call `s3.getBucketLocation` twice: 1) without `ExpectedBucketOwner` and 2) with `ExpectedBucketOwner` set to the target account. If (1) results in a success, but (2) results in `NO_ACCESS`, we determine we have access, but the bucket doesn't belong to the target account.

In addition, add an option to instruct asset publishing to fail if indeed a foreign bucket is detected.