chore: create aliases for all SDK input and output types (#124)
To avoid exposing SDK v3 types in the public interface, we are copy-pasting all of them. If in the future we have to migrate to a v4 of the SDK, consumers of this package will be shielded from breaking changes, as we can adapt the type as an implementation detail.