fix: force SHA256 checksums (#153)
By default, the S3 client will use MD5 checksums. But in FIPS
environments, those MD5 checksums will not work because Node will not
have access to an M5 digest.

Force the usage of SHA256 for these checksums instead. In SDKv3, this is
as simple as configuring the algorithm to use, and the SDK itself will
do the rest.

Relates to https://github.com/aws/aws-cdk/issues/31926