chore: tests are not isolated
The `upload external asset correctly` test was asserting that
`PutObject` would be called twice, but it was only called twice
because a test run prior to it leaked a `PubObject` invocation that
was counted towards the count of this test.

If the test was run by itself, the `PutObject` count was `1`, not `2`,
and the test would fail.