fix: does not work with newer SDK versions (#299)
With newer SDK versions, our copy of the AWS types is incorrect.

We have this copy so that we can abstract over the specific SDK used
(even if we move to SDKv4 in the future) and only care about the input
and output parameters. But the underlying SDKv3 has added a new value to
the `ChecksumAlgorithm` enum, and now we don't properly typecheck
anymore because the SDK output can set an enum to a value we don't
support.

The proper type for an enum that is as an output is not actually the
enum, but `string`: the service team can decide to add new values to the
enum at any point in time, and our restrictive `enum` type would be
lying immediately and consumers would be unprepared to handle it.

That's why I'm introducing machinery to have different types depending
on whether we are using the checksum algorithm as an input or output: as
an input, it can be the enum. As an output, it must be `string`.

The `InOut` type was actually not necessary since it looks like we're
not using `_Object` as an input anywhere; I'm still keeping it in
because it will be useful machinery to have in the back pocket next time
this happens.

Also remove an unused variable which leads to a compiler error in a
stricter build environment.