feat: emit all shell logging to publishProgressListener (#295)
Closes #196 

Removes quiet flag and also ensures all events are routed to the
progressListener.

removeD the existing `logger` callback and replaces it with
`shellEventPublisher` which will route the message back to the
configured progressListener if configured. The default behavior is
`stdio` for the new `subprocessOutputDestination` option, which
essentially is the old behavior. Where shells write directly to
stdout/stderr

---------

Co-authored-by: Rico Hermans <rix0rrr@gmail.com>