fix: docker build attestations break cdk-assets (400 Bad Request) (#342)
There are various issues in cdk that can be traced back to attestations
in docker:

https://github.com/aws/aws-cdk/issues/30258
https://github.com/aws/aws-cdk/issues/31549
https://github.com/aws/aws-cdk/issues/33264

cdk-assets cannot work with docker containerd because it will attempt to
upload multiple files to the same hash in ECR, and our ECR repository is
immutable (by requirement). docker recently changed their default to
turn on containerd which causes this issue to skyrocket.

the hotfix here is to add an environment variable when calling `docker`
so that the attestation file is not added to the manifest. we can later
look into adding support for including
[provenance](https://docs.docker.com/build/metadata/attestations/slsa-provenance/)
attestations if there is need for it.

i've chosen to fix this via environment variable instead of as a command
option `--provenance=false` because we must keep docker replacements in
mind, and at least finch [does
not](https://runfinch.com/docs/cli-reference/finch_build/) have a
`provenance` property at the moment.

in addition to this unit test that shows the env variable exists when
`docker build` is called, i have also ensured that this solves the issue
in my local setup + symlinked `cdk-assets`..

(cherry picked from commit 8bdea135b8c3ee2f28f0c18f8d8ec912442c9025)

# Conflicts:
#	lib/private/docker.ts
#	test/private/docker.test.ts